refact: Migrate UUID commands to command classes
